refactor: put shared logic into filter-builder-base.ts script
